+++ src/Order/Models/OrderStatus.php
@@ -19,14 +19,17 @@ use Vanilo\Order\Contracts\OrderStatus as OrderStatusContract;
 
 /**
  * @method static OrderStatus PENDING();
+ * @method static OrderStatus PROCESSING();
  * @method static OrderStatus COMPLETED();
  * @method static OrderStatus CANCELLED();
  *
  * @method bool isPending()
+ * @method bool isProcessing()
  * @method bool isCompleted()
  * @method bool isCancelled()
  *
  * @property-read bool $is_pending
+ * @property-read bool $is_processing
  * @property-read bool $is_completed
  * @property-read bool $is_cancelled
  */
@@ -38,6 +41,9 @@ class OrderStatus extends Enum implements OrderStatusContract
      * Pending orders are brand new orders that have not been processed yet.
      */
     public const PENDING = 'pending';
+
+    /** Order is being packed, manufactured, or partially fulfilled */
+    public const PROCESSING = 'processing';
 
     /**
      * Orders fulfilled completely.
@@ -52,7 +58,7 @@ class OrderStatus extends Enum implements OrderStatusContract
     // $labels static property needs to be defined
     public static $labels = [];
 
-    protected static $openStatuses = [self::PENDING];
+    protected static $openStatuses = [self::PENDING, self::PROCESSING];
 
     public function isOpen(): bool
     {
@@ -68,6 +74,7 @@ class OrderStatus extends Enum implements OrderStatusContract
     {
         static::$labels = [
             self::PENDING => __('Pending'),
+            self::PROCESSING => __('Processing'),
             self::COMPLETED => __('Completed'),
             self::CANCELLED => __('Cancelled')
         ];