+++ Changelog.md
@@ -6,6 +6,7 @@
 ##### 2023-XX-YY
 
 - Added master product & variant (Foundation) to the morph map
+- Added the `processing` order status enum value
 - Added Zone support to shipping methods (optional)
 - Changed minimum Address module requirement to v2.5